requirements.txt
asgiref==3.5.2
Django==3.2
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
